refactor(handler): move run lifecycle handlers to runs_lifecycle.go (#1022)
Pure code motion: relocate the run create/dispatch/get/messages/input HTTP
handlers and the guard-role-excluded notification glue out of workers.go into
the new runs_lifecycle.go. --color-moved clean; no exported-API or behaviour
change. workers.go retains only the worker handlers and DTO builders; its
import set is pruned to what remains.
